# src/test/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "HRD Room Service API",
    "description" : "REST API for HRD room service operations, including authentication, academic years, classrooms, courses, and enrollments. Obtain a JWT from POST /api/v4/auth/login; send it as Bearer authentication on protected routes.",
    "contact" : {
      "name" : "HRD Room Service Team",
      "email" : "penhseyha4980@gmail.com"
    }
  },
  "paths" : {
...
    "/api/v4/academic-years/{academicYearId}/archive" : {
      "put" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "academicYearId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/academic-years/{academicYearId}/activate" : {
      "put" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "activate",
        "parameters" : [ {
          "name" : "academicYearId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/academic-years" : {
      "get" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "includeArchived",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AcademicYearRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v4/academic-years/{academicYearId}" : {
      "get" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "byId_3",
        "parameters" : [ {
          "name" : "academicYearId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v4/academic-years/active" : {
      "get" : {
        "tags" : [ "academic-year-controller" ],
        "operationId" : "active",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAcademicYearResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AcademicYearResponse" : {
        "type" : "object",
        "properties" : {
          "academicYearId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "generation" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseAcademicYearResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AcademicYearResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
      "AcademicYearRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "generation" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "endDate", "name", "startDate" ]
      },
...
      "ApiResponsePageResponseAcademicYearResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseAcademicYearResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "errorCode" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
      "PageResponseAcademicYearResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AcademicYearResponse"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}